feat: update background images
