+++ apps/webapp/src/script/components/calling/FullscreenVideoCall.tsx
@@ -36,8 +36,8 @@ import {
 import {WebAppEvents} from '@wireapp/webapp-events';
 
 import {useAppNotification} from 'Components/AppNotification/AppNotification';
-import {PerformancePanel} from 'Components/calling/PerformancePanel/PerformancePanel';
 import {useCallAlertState} from 'Components/calling/useCallAlertState';
+import {VideoBackgroundPerformancePanel} from 'Components/calling/VideoControls/videoBackgroundPerformancePanel/videoBackgroundPerformancePanel';
 import {ConversationClassifiedBar} from 'Components/ClassifiedBar/ClassifiedBar';
 import * as Icon from 'Components/Icon';
 import {ModalComponent} from 'Components/Modals/ModalComponent';
@@ -194,7 +194,7 @@ const FullscreenVideoCall = ({
   const openPopup = () => callingRepository.setViewModeDetached();
 
   const [isParticipantsListOpen, toggleParticipantsList] = useToggleState(false);
-  const [isBackgroundSidebarOpen, setIsBackgroundSidebarOpen] = useState(false);
+  const [isBackgroundSidebarOpen, setIsBackgroundSidebarOpen] = useState(true);
 
   const callNotification = useAppNotification({
     activeWindow: viewMode === CallingViewMode.DETACHED_WINDOW ? detachedWindow! : window,
@@ -573,7 +573,7 @@ const FullscreenVideoCall = ({
           </>
         )}
       </ModalComponent>
-      <PerformancePanel backgroundEffectsHandler={backgroundEffectsHandler} />
+      <VideoBackgroundPerformancePanel backgroundEffectsHandler={backgroundEffectsHandler} />
     </div>
   );
 };
